ChatBot1/requirements.txt
Flask
google-generativeai
torch
transformers
